Add demo sidebar
Since we want a sidebar on article pages, I modified the PostTemplate. I
also tried to emulate the component structure that is being used by
creating a Sidebar component similar to the Article component.

I even mostly copy-pasted the style of the Article component. This
doesn't really make sense, specially in the case of the CSS max-width,
which will not be the same for the sidebar.

I also didn't bother to build a proper wrapper component for Article and
Sidebar in PostTemplate and just added the grid styling inline.